% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCAA_CNIG}
\alias{CCAA_CNIG}
\title{CCAA_CNIG.rda}
\format{Es Spatial Polygon Dataframe, pero su df tiene 18 y 7 columnas
\itemize{
  \item NombreCCAA: nombre CCAA
  \item INECodCCAA: codigo CCAA
  \item INENumCCAA: numero CCAA
  \item Area:
  \item Perimeter:
  \item X_Centroid:
  \item Y_Centroid:
}}
\source{
\url{https://www.cnig.es/}
}
\usage{
CCAA_CNIG
}
\description{
Shapefiles de CCAA de CNIG
https://www.cnig.es/
}
\keyword{datasets}
