% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CNIG_CCAA_sf}
\alias{CNIG_CCAA_sf}
\title{CNIG_CCAA_sf.rda}
\format{Es simple feauture 18 rows y 8 columnas
\itemize{
  \item NombreCCAA: nombre CCAA
  \item INECodCCAA: codigo CCAA
  \item INENumCCAA: numero CCAA
  \item Area:
  \item Perimeter:
  \item X_Centroid:
  \item Y_Centroid:
}}
\source{
\url{https://www.cnig.es/}
}
\usage{
CNIG_CCAA_sf
}
\description{
Shapefiles de CCAA de CNIG
https://www.cnig.es/
}
\details{
Simple feature collection with 18 features and 7 fields
geometry type:  MULTIPOLYGON
dimension:      XY
bbox:           xmin: 1547225 ymin: 941709.3 xmax: 3833655 ymax: 2468216
epsg (SRID):    NA
proj4string:    +proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +units=m +no_defs
}
\keyword{datasets}
