% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CNIG_municipios_sf}
\alias{CNIG_municipios_sf}
\title{CNIG_municipios_sf.rds}
\format{Es Ssimple feature 8116 rows y 13 columnas (+ la geometry)
pero resulta que en el fichero del INE tengo 8125 municipios
\itemize{
  \item NombreMuni: nombre del municipio
  \item INECodMuni: codigo del municipio [tiene 5 dígitos; cod_prov+cod_muni]
  \item INENumMuni: número del municipio
  \item NombreProv:
  \item INECodProv:
  \item INENumProv:
  \item NombreCCAA:
  \item INECodCCAA:
  \item INENumCCAA:
  \item Area:
  \item Perimeter:
  \item X_Centroid:
  \item Y_Centroid:
}}
\source{
\url{https://www.cnig.es/}
}
\usage{
CNIG_municipios_sf
}
\description{
Shapefiles de municipios de CNIG
https://www.cnig.es/
}
\details{
Reading layer Municipios2011_ETRS89_LAEA from data Municipios2011_ETRS89_LAEA.shp using driver ESRI Shapefile
Reading layer `Municipios2011_ETRS89_LAEA' from data source Municipios2011_ETRS89_LAEA.shp using driver `ESRI Shapefile'
Simple feature collection with 8116 features and 13 fields
geometry type:  MULTIPOLYGON
dimension:      XY
bbox:           xmin: 1547225 ymin: 941709.3 xmax: 3833655 ymax: 2468216
epsg (SRID):    NA
proj4string:    +proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +units=m +no_defs
}
\keyword{datasets}
