% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CNIG_provincias_sf}
\alias{CNIG_provincias_sf}
\title{CNIG_provincias_sf.rda}
\format{Es simple features  52 rows y 11 columnas
\itemize{
  \item NombreProv: nombre provincia
  \item INECodProv: codigo provincia
  \item INENumProv: numero provincia
  \item NombreCCAA:
  \item INECodCCAA:
  \item INENumCCAA:
  \item Area:
  \item Perimeter:
  \item X_Centroid:
  \item Y_Centroid:
}}
\source{
\url{https://www.cnig.es/}
}
\usage{
CNIG_provincias_sf
}
\description{
Shapefiles de PROVINCIAS de CNIG
https://www.cnig.es/
}
\details{
Reading layer `Provincias2011_ETRS89_LAEA' from data source Provincias2011_ETRS89_LAEA.shp using driver ESRI Shapefile
Simple feature collection with 52 features and 10 fields
geometry type:  MULTIPOLYGON
dimension:      XY
bbox:           xmin: 1547225 ymin: 941709.3 xmax: 3833655 ymax: 2468216
epsg (SRID):    NA
proj4string:    +proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +units=m +no_defs
}
\keyword{datasets}
