% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod_INE_muni_pjp}
\alias{cod_INE_muni_pjp}
\title{cod_INE_muni_pjp.rda}
\format{A data frame with 8125 rows and 7 variables:
\itemize{
  \item INECodMuni: Código de provincia + Código municipio. Es el que usa el INE muchas veces
  \item NOMBRE: Nombre del municipio
  \item CPRO: Código de la provincia
  \item DC: Código de control
  \item CODAUTO: Código de CCAA
  \item CCAA: Nombre de la CCAA
  \item PROVINCIA: Nombre de la provincia
}}
\source{
\url{http://www.ine.es}
}
\usage{
cod_INE_muni_pjp
}
\description{
Codigos del INE para municipios españoles a 1 enero 2016
tb tiene los codigos de provincia y CCAA
los he unido yo asi:
}
\details{
cod_municipios_16 <- cod_municipios_16 %>% mutate(INECodMuni = paste0(CPRO, CMUN))
cod_municipios_16 <- cod_municipios_16 %>% select(INECodMuni, NOMBRE, CPRO, DC)
cod_INE_muni_pjp <-  left_join(cod_municipios_16, cod_provincias, by = "CPRO")
}
\keyword{datasets}
