% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_PEN_2018}
\alias{df_PEN_2018}
\title{df_PEN_2018.rda}
\format{Es una tibble, con unas 483 rows y 35 variables
\itemize{
  \item code_PEN: es el código de la operacion estadistica en el PEN (este es el GOOD)
  \item op_estadistica: el nombre de la operacion estadistica
  \item code_IOE: código INE del "inventario de operaciones estadisticas" Es otro codigo al del PEN
  \item code_PEN_filled: code_PEN rellenando los NA accordingly (pero no lo uses salvo q realmnete lo quieras)
  \item ppto_total: coste de la encuenta (solo 2 operaciones tienen 2 organismos financiadores)
  \item id_sector: id sector o tema
  \item sector: sector o tema
  \item id_ministerio: cada operacion esta adscrita a un Ministerio
  \item Ministerio: nombre del Ministerio
  \item org_ppto1: siglas del organismo que financia la operacion estadistica
  \item org_involu1: el primero de la lista de organismo involucrados en la encuesta (llega a ver hasta 12 organismos)
}}
\usage{
df_PEN_2018
}
\description{
Hack del Plan Estadistico Nacional 2018
Los detalles en "./data-raw/script_hack_PEN_2018_v1.R"
Lo q si t has de acordar es que en 2018 habia 407 operaciones estadisticas
PERO q como aparte del code_PEN hay otro codigo (code_IOE)
pues por eso hay 483 filas pero solo 400 operaciones del PEN
o sea, hay 76 con la v. code_PEN = NA
}
\details{
Hay mas variables de las que muestro (generalmente de organismos involucrados)
}
\keyword{datasets}
