% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esp_adm0}
\alias{esp_adm0}
\title{esp_adm0.rda
Shapefiles de GADM database of Global Administrative Areas: COUNTRY level
http://www.gadm.org/country}
\format{An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 68 columns.}
\source{
\url{http://www.gadm.org/country}
}
\usage{
esp_adm0
}
\description{
su df tiene 68 variables !! todas referidada a SPAIN (nombre en frances...)
}
\keyword{datasets}
