% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esp_adm1}
\alias{esp_adm1}
\title{esp_adm1.rda}
\format{Es Spatial Polygon, pero su df tiene 18 rows (CCAA) y 15 variables ellas
\itemize{
  \item CODAUTO: Código de CCAA (original del INE, fichero cod_provincias.rda)
  \item CCAA: Nombre de la CCAA (original del INE, fichero cod_provincias.rda)
}}
\source{
\url{http://www.gadm.org/country}
}
\usage{
esp_adm1
}
\description{
Shapefiles de GADM database of Global Administrative Areas: CCAA
http://www.gadm.org/country
}
\details{
Tiene un PB: Ceuta y Melilla estaban juntas, asi que le he asignado el Código
y el nombre de Ceuta; osea, el nombre "Ceuta" está asignado a un polygon que recoge
los limites de Ceuta y de melilla

tb tiene los nombres y codigos  de las CCAA del INE xq los he añadido yo asín:
los he unido yo asi:

esp_adm1 <- readRDS(file = "./data-raw/ESP_adm1.rds")  #- shape GADM de ESPAÑA pais
load("./data/cod_provincias.rda")
esp_adm1@data$CCAA <- esp_adm1@data$NAME_1 #- duplico NAME_2 para arrglarlo con los codigos INE
esp_adm1@data$CCAA[esp_adm1@data$CCAA == "Principado de Asturias"] <- "Asturias, Principado de"
esp_adm1@data$CCAA[esp_adm1@data$CCAA == "Islas Baleares"] <- "Balears, Illes"
esp_adm1@data$CCAA[esp_adm1@data$CCAA == "Islas Canarias"] <- "Canarias"
esp_adm1@data$CCAA[esp_adm1@data$CCAA == "Comunidad Valenciana"] <- "Comunitat Valenciana"
esp_adm1@data$CCAA[esp_adm1@data$CCAA == "Comunidad de Madrid"] <- "Madrid, Comunidad de"
esp_adm1@data$CCAA[esp_adm1@data$CCAA == "Región de Murcia"] <- "Murcia, Región de"
esp_adm1@data$CCAA[esp_adm1@data$CCAA == "Comunidad Foral de Navarra"] <- "Navarra, Comunidad Foral de"
esp_adm1@data$CCAA[esp_adm1@data$CCAA == "La Rioja"] <- "Rioja, La"
esp_adm1@data$CCAA[esp_adm1@data$CCAA == "Ceuta y Melilla"] <- "Ceuta"  #- MAL, xq son los ines de Ceuta y Melilla
}
\keyword{datasets}
