% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esp_adm2}
\alias{esp_adm2}
\title{esp_adm2.rda}
\format{Es Spatial Polygon, pero su df tiene 18 rows (CCAA) y 15 variables ellas
\itemize{
  \item CPRO: Código de la provincia (original del INE, fichero cod_provincias.rda)
  \item PROVINCIA: Nombre de la provincia  (original del INE, fichero cod_provincias.rda)
  \item CODAUTO: Código de CCAA (original del INE, fichero cod_provincias.rda)
  \item CCAA: Nombre de la CCAA (original del INE, fichero cod_provincias.rda)
}}
\source{
\url{http://www.gadm.org/country}
}
\usage{
esp_adm2
}
\description{
Shapefiles de GADM database of Global Administrative Areas: PROVINCIAS
http://www.gadm.org/country
}
\details{
tb tiene los nombres y codigos de CCAA y PROVINCIAS del INE xq los he añadido yo asín:
los he unido yo asi:
}
\keyword{datasets}
