% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esp_adm3}
\alias{esp_adm3}
\title{esp_adm3.rda}
\format{Es Spatial Polygon, pero su df tiene 369 rows y 16 columnas
¿xq hay 369 comarcas, si realmente muchos tienen NAME_3 = %na? Pues xq SI, molan las comarcas}
\source{
\url{http://www.gadm.org/country}
}
\usage{
esp_adm3
}
\description{
Shapefiles de GADM database of Global Administrative Areas: COMARCAS
http://www.gadm.org/country
}
\keyword{datasets}
