% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esp_adm4}
\alias{esp_adm4}
\title{esp_adm4.rda}
\format{Es Spatial Polygon, pero su df tiene 8302 y 17 columnas
pero resulta que en el fichero del INE tengo 8125 municipios}
\source{
\url{http://www.gadm.org/country}
}
\usage{
esp_adm4
}
\description{
Shapefiles de GADM database of Global Administrative Areas: MUNICIPIOS
http://www.gadm.org/country
}
\keyword{datasets}
