% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Esta función solo saluda a mi blog
Es q quería q se modificase el fichero NAMESPACE}
\usage{
hello()
}
\description{
Esta función solo saluda a mi blog
Es q quería q se modificase el fichero NAMESPACE
}
