% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{listado_organismos_PEN}
\alias{listado_organismos_PEN}
\title{listado_organismos_PEN.rda}
\format{Es una tibble, con unas 35 rows y 2 variables
\itemize{
  \item sigla_org: siglas del organismo (BANESP)
  \item name_org: nombre del organismo (Banco de España)
}}
\usage{
listado_organismos_PEN
}
\description{
Listado de organismos en el Plan Estadistico Nacional (estan en un apendice)
Los detalles en "./data-raw/script_hack_PEN_2018_v1.R"
}
\keyword{datasets}
