% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipios_CNIG}
\alias{municipios_CNIG}
\title{municipios_CNIG.rda}
\format{Es Spatial Polygon Dataframe, pero su df tiene 8116 y 13 columnas
pero resulta que en el fichero del INE tengo 8125 municipios
\itemize{
  \item NombreMuni: nombre del municipio
  \item INECodMuni: codigo del municipio [tiene 5 dígitos; cod_prov+cod_muni]
  \item INENumMuni: número del municipio
  \item NombreProv:
  \item INECodProv:
  \item INENumProv:
  \item NombreCCAA:
  \item INECodCCAA:
  \item INENumCCAA:
  \item Area:
  \item Perimeter:
  \item X_Centroid:
  \item Y_Centroid:
}}
\source{
\url{https://www.cnig.es/}
}
\usage{
municipios_CNIG
}
\description{
Shapefiles de municipios de CNIG
https://www.cnig.es/
}
\keyword{datasets}
