% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypkgfordata.R
\docType{package}
\name{mypkgfordata}
\alias{mypkgfordata}
\alias{mypkgfordata-package}
\title{This file documents the package
mypkgfordata: datasets used in my blog}
\description{
This package have datasets used in the posts
of mi blog
R & Flowers II   at
rflowers2.netlify.com
}
