% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{padron_15_total_x_nac}
\alias{padron_15_total_x_nac}
\title{padron_15_total_x_nac.rda}
\format{Es una tibble, con 114 rows y 4 variables
\itemize{
  \item pais_nacimiento:
  \item municipios: en este df esta v. es inutil xq los datos son para el total nacional, pero lo dejo
  \item sexo: [Total, Hombre, Mujer]
  \item value: nº de personas (integer)
}}
\usage{
padron_15_total_x_nac
}
\description{
Datos de poblacion (por nacionalidad) para el total de Spain
datos procedentes del Padron de 2015
}
\keyword{datasets}
