% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{padron_15_x_edad_3g}
\alias{padron_15_x_edad_3g}
\title{padron_15_x_edad_3g.rda}
\format{Es una tibble, con 292.284 rows y 6 variables
\itemize{
  \item edad_3g: grupo de edad (3+1 grupo)(Total [0-16[, [16-65[  y 65 o mas
  \item nacionalidad: Total, españoles, extranjeros
  \item INECodMuni: Código de provincia + Código municipio. Es el que usa el INE muchas veces
  \item NombreMuni: nombre del municipio
  \item sexo: [Total, Hombre, Mujer]
  \item value: nº de personas (integer)
}}
\usage{
padron_15_x_edad_3g
}
\description{
Datos de poblacion (3 grupos de edad + el Total) para cada MUNICIPIO
datos procedentes del Padron de 2015
las v. INECodMuni y NombreMuni salen de separar la v. "municipios"
df_x <- df_x %>% separate(municipios, into= c("INECodMuni", "NombreMuni"), by = "-", extra = "merge")
}
\keyword{datasets}
