% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{padron_15_x_edad_5}
\alias{padron_15_x_edad_5}
\title{padron_15_x_edad_5.rda}
\format{Es una tibble, con 535.854 rows y 5 variables
\itemize{
  \item edad_5: grupo de edad (quinquenios)
  \item INECodMuni: Código de provincia + Código municipio. Es el que usa el INE muchas veces
  \item NombreMuni: nombre del municipio
  \item sexo: [Total, Hombre, Mujer]
  \item value: nº de personas (integer)
}}
\usage{
padron_15_x_edad_5
}
\description{
Datos de poblacion (por quinquenios de edad) para cada MUNICIPIO
datos procedentes del Padron de 2015
las v. INECodMuni y NombreMuni salen de separar la v. "municipios"
df_x <- df_x %>% separate(municipios, into= c("INECodMuni", "NombreMuni"), by = "-", extra = "merge")
}
\keyword{datasets}
