% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{padron_15_x_nac}
\alias{padron_15_x_nac}
\title{padron_15_x_nac.rda}
\format{Es una tibble, con 925.566 rows y 5 variables
\itemize{
  \item pais_nacimiento:
  \item INECodMuni: Código de provincia + Código municipio. Es el que usa el INE muchas veces
  \item NombreMuni: nombre del municipio
  \item sexo: [Total, Hombre, Mujer]
  \item value: nº de personas (integer)
}}
\usage{
padron_15_x_nac
}
\description{
Datos de poblacion (por nacionalidad) para cada MUNICIPIO
datos procedentes del Padron de 2015
las v. INECodMuni y NombreMuni salen de separar la v. "municipios"
df_x <- df_x %>% separate(municipios, into= c("INECodMuni", "NombreMuni"), by = "-", extra = "merge")
}
\keyword{datasets}
