% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{padron_15_x_nac_res}
\alias{padron_15_x_nac_res}
\title{padron_15_x_nac_res.rda}
\format{Es una tibble, con 219.213 rows y 5 variables
\itemize{
  \item res_nac: 8 gruppos + Total (x residencia y nacimineto)p.ej:Misma Comunidad Autónoma. Misma Provincia
  \item INECodMuni: Código de provincia + Código municipio. Es el que usa el INE muchas veces
  \item NombreMuni: nombre del municipio
  \item sexo: [Total, Hombre, Mujer]
  \item value: nº de personas (integer)
}}
\usage{
padron_15_x_nac_res
}
\description{
Datos de poblacion por relacion entre residencia y nacimineto (9 grupos)
datos procedentes del Padron de 2015
las v. INECodMuni y NombreMuni salen de separar la v. "municipios"
df_x <- df_x %>% separate(municipios, into= c("INECodMuni", "NombreMuni"), by = "-", extra = "merge")
}
\keyword{datasets}
