% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provincias_CNIG}
\alias{provincias_CNIG}
\title{provincias_CNIG.rda}
\format{Es Spatial Polygon Dataframe, pero su df tiene 52 y 10 columnas
\itemize{
  \item NombreProv: nombre provincia
  \item INECodProv: codigo provincia
  \item INENumProv: numero provincia
  \item NombreCCAA:
  \item INECodCCAA:
  \item INENumCCAA:
  \item Area:
  \item Perimeter:
  \item X_Centroid:
  \item Y_Centroid:
}}
\source{
\url{https://www.cnig.es/}
}
\usage{
provincias_CNIG
}
\description{
Shapefiles de PROVINCIAS de CNIG
https://www.cnig.es/
}
\keyword{datasets}
